+++ src/allotropy/parsers/agilent_gen5/agilent_gen5_structure.py
@@ -6,7 +6,7 @@ from collections import defaultdict
 from dataclasses import dataclass
 import re
 
-from allotropy.allotrope.models.plate_reader_benchling_2023_09_plate_reader import (
+from allotropy.allotrope.models.adm.plate_reader.benchling._2023._09.plate_reader import (
     ScanPositionSettingPlateReader,
 )
 from allotropy.allotrope.models.shared.definitions.definitions import JsonFloat