+++ src/allotropy/parsers/beckman_pharmspec/beckman_pharmspec_parser.py
@@ -3,7 +3,7 @@ from typing import Any, Optional, Union
 
 import pandas as pd
 
-from allotropy.allotrope.models.light_obscuration_benchling_2023_12_light_obscuration import (
+from allotropy.allotrope.models.adm.light_obscuration.benchling._2023._12.light_obscuration import (
     CalculatedDataDocumentItem,
     DataProcessingDocument,
     DataSourceDocumentItem,