+++ src/allotropy/parsers/qiacuity_dpcr/qiacuity_dpcr_parser.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import pandas as pd
 
-from allotropy.allotrope.models.pcr_benchling_2023_09_dpcr import (
+from allotropy.allotrope.models.adm.pcr.benchling._2023._09.dpcr import (
     DataProcessingDocument,
     DataSystemDocument,
     DeviceControlAggregateDocument,